// generated/nextjs/search-index/types.ts
export interface SearchIndexStorage {
  readonly get: (relation: string, key: string) => Promise<Record<string, unknown> | null>;
  readonly put: (relation: string, key: string, value: Record<string, unknown>) => Promise<void>;
  readonly delete: (relation: string, key: string) => Promise<boolean>;
  readonly find: (relation: string, filter?: Record<string, unknown>) => Promise<readonly Record<string, unknown>[]>;
}

export interface SearchIndexCreateIndexInput {
  readonly index: string;
  readonly config: string;
}

export interface SearchIndexCreateIndexOutputOk {
  readonly variant: 'ok';
  readonly index: string;
}

export interface SearchIndexCreateIndexOutputExists {
  readonly variant: 'exists';
  readonly index: string;
}

export type SearchIndexCreateIndexOutput = SearchIndexCreateIndexOutputOk | SearchIndexCreateIndexOutputExists;

export interface SearchIndexIndexItemInput {
  readonly index: string;
  readonly item: string;
  readonly data: string;
}

export interface SearchIndexIndexItemOutputOk {
  readonly variant: 'ok';
  readonly index: string;
}

export interface SearchIndexIndexItemOutputNotfound {
  readonly variant: 'notfound';
  readonly index: string;
}

export type SearchIndexIndexItemOutput = SearchIndexIndexItemOutputOk | SearchIndexIndexItemOutputNotfound;

export interface SearchIndexRemoveItemInput {
  readonly index: string;
  readonly item: string;
}

export interface SearchIndexRemoveItemOutputOk {
  readonly variant: 'ok';
  readonly index: string;
}

export interface SearchIndexRemoveItemOutputNotfound {
  readonly variant: 'notfound';
  readonly index: string;
}

export type SearchIndexRemoveItemOutput = SearchIndexRemoveItemOutputOk | SearchIndexRemoveItemOutputNotfound;

export interface SearchIndexSearchInput {
  readonly index: string;
  readonly query: string;
}

export interface SearchIndexSearchOutputOk {
  readonly variant: 'ok';
  readonly results: string;
}

export interface SearchIndexSearchOutputNotfound {
  readonly variant: 'notfound';
  readonly index: string;
}

export type SearchIndexSearchOutput = SearchIndexSearchOutputOk | SearchIndexSearchOutputNotfound;

export interface SearchIndexAddProcessorInput {
  readonly index: string;
  readonly processor: string;
}

export interface SearchIndexAddProcessorOutputOk {
  readonly variant: 'ok';
  readonly index: string;
}

export interface SearchIndexAddProcessorOutputNotfound {
  readonly variant: 'notfound';
  readonly index: string;
}

export type SearchIndexAddProcessorOutput = SearchIndexAddProcessorOutputOk | SearchIndexAddProcessorOutputNotfound;

export interface SearchIndexReindexInput {
  readonly index: string;
}

export interface SearchIndexReindexOutputOk {
  readonly variant: 'ok';
  readonly count: number;
}

export interface SearchIndexReindexOutputNotfound {
  readonly variant: 'notfound';
  readonly index: string;
}

export type SearchIndexReindexOutput = SearchIndexReindexOutputOk | SearchIndexReindexOutputNotfound;